# ENH: stats.rankdata: vectorize calculation