#### Reference issue
gh-19770
gh-19749

#### What does this implement/fix?
`stats.rankdata` implemented `axis` by looping over slices; this vectorizes the calculation. It will be useful for the referenced PRs.

#### Additional information
This PR version tends to be a little faster for both small and large 1D arrays and considerably faster with multidimensional arrays over an axis. It can be a little slower for `nan_policy='propagate'` right now because it calculates the ranks in all slices before replacing slices with NaNs, but there is a simple fix for that if we think it's important.

This will immediately speed up `rankdata` and many hypothesis tests that depend on it. It also opens the door to further improvements in these hypothesis tests because `_rankdata` can also return tie counts, which many of these tests need.

<details>

```
import numpy as np
from scipy import stats
rng = np.random.default_rng(435984535)

# shape = (100)
shape = (100, 100)
axis = -1

x = rng.integers(0, 5, size=shape)
# %timeit stats.rankdata(x, axis=-1)
# pr   1d: 31.3 µs ± 442 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# main 1d: 37.2 µs ± 233 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# pr   2d: 369 µs ± 4.1 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# main 2d: 2.58 ms ± 8.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

y = x.copy().astype(float)
i = rng.random(size=shape)
y[i < 0.3] = np.nan
# %timeit stats.rankdata(y, axis=-1, nan_policy='omit')
# pr   1d: 40.1 µs ± 122 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# main 1d: 46.9 µs ± 586 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# pr   2d: 526 µs ± 6.29 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# main 2d: 3.59 ms ± 28.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

z = rng.normal(size=shape)
# %timeit stats.rankdata(z, axis=-1)
# pr   1d: 38.9 µs ± 57.7 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# main 1d: 44.8 µs ± 104 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
# pr   2d: 604 µs ± 10.7 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# main 2d: 3.47 ms ± 23.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
</details>