@mdhaber 2023-12-29
@AtsushiSakai we've worked on `rankdata` together before. Would you be willing to review this?