@mdhaber 2024-01-01
Thanks @AtsushiSakai. Were there other comments, or should I make these changes and re-run CI?  I suppose in the meantime, I'll use `if`/`elif` instead of `match`/`case`.