@mdhaber 2024-01-01
Thanks @AtsushiSakai. I implemented the recommendation. Do you think this is ready for merge, or does it need another review? 
When it's done, I can use it in the linked `wilcoxon` and `mannwhitneyu` PRs. Are you interested in those, BTW?