@AtsushiSakai 2024-01-02
> Are you interested in those, BTW?

i’m not familiar with these functions, but yeah. I can help to review the PRs．